+++ dataflow/src/main/java/org/checkerframework/dataflow/cfg/builder/CFGTranslationPhaseOne.java
@@ -1430,7 +1430,7 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
     }
 
     List<Node> arguments;
-    if (TreeUtils.isEnumSuper(tree)) {
+    if (TreeUtils.isEnumSuperCall(tree)) {
       // Don't convert arguments for enum super calls.  The AST contains no actual arguments, while
       // the method element expects two arguments, leading to an exception in convertCallArguments.
       // Since no actual arguments are present in the AST that is being checked, it shouldn't cause