+++ docs/CHANGELOG.md
@@ -13,6 +13,8 @@ Renamed `TreeUtils.isEnumSuper` to `isEnumSuperCall`.
 
 **Closed issues:**
 
+#5390, #5399, #5390.
+
 
 Version 3.27.0 (November 1, 2022)
 ---------------------------------